RM6048: accounts in Accounting Balance changes
Customer, supplier and employee accounts in Accounting Balance (in
Partner) are now null by default. The default accounts used will be from
the company. But if the account values are set in Partner, these values
will be used instead.
When generating accounting from expenses, if the employee accounting
balance line is missing, we now create it (instead of showing an error).