RM7223: Moved default ebics signatory user
Ebics signatory user moved from account configuration to ebics partner.
In bank order, replaced ebicsUser by signatoryEbicsUser. The field
signatoryUser is now hidden and the field signatoryEbicsUser is shown
instead.
Fill signatoryEbicsUser with the default from ebics partner. It is
possible to change it, but the ebics user must be linked with the bank
details through the ebics partner.