RM7222: filter on BankDetails in BankOrderLine
In a line of a bank order, the bank details domain is now filtered by the
list of bank details in ebics partner.
In bank order line, we also fill the bank details on partner/company
change if the default bak details in the partner is in the ebics partner
list.
Fixed the visibility of partner/company fields in the bank order line.