SequenceService: be threadsafe and singleton
SequenceService wasn't declared as a Singleton which seems rather strange
(but this implied that issue #613 wasn't an issue except if sequence
service was to be used in @SingletonClass binding it in their constructor).
Fix this and make it thread safe by dropping any mutable members. Calls
have been adjusted when needed.
This means that we've now one single instance of SequenceService instead
of having one instance per injection (which is a LOT less instances).