RM9214: Change display conditions, export format and rearrangement of main panel
The import button now appears only in "In progress" and "Completed"
state.
The title of the import button now displays the expected format (CSV).
The "fill inventory with" panel disappears in "Planned" state.
Date and location are now in read-only state when Inventory state is
"Planned".
There is no more empty space on the main panel right side.
CSV exportation now includes headers and the separator is changed to
";".
Add Try-Catch statement to controller to trace any generated anomaly.