db: add mechanism allowing to register lifecycle listeners
Add an Hibernate integrator that automatically scan for classes
implementing a given interface which provides hooks to be notified of
various persistence events, without modifying entities. This could allow
to add additional consistency checks.

Beware that such listeners are bound to the same limitations as native
entity listeners, meaning that they cannot use entity manager or make
queries.

This should be moved to axelor core if of any use.