prestashop: fix potential issue due to transactions
Export process (and I guess import one too but haven't started to look
at it right now) was using last batch's end date as a reference date to
know if an entity has been modified since last run. This is wrong as
entities could be updated during the run and thus be missed on the next
one.
Note that due to transactions interlacing, we could still miss updates,
on an heavily loaded platform, if an entity is updated *before* the batch
(setting updatedOn) but its transaction is commited during it, it will
be missed.