prestashop: Batch process method shouldn't be transactional
In this specific case, we're playing with a distributed architecture.
This mean that we cannot just rollback local transaction an assume
everything will go well. In fact, you'll end up with duplicate entities
(on prestashop side) all around (except for products and currencies
which have an unique code).
Removing @Transactional on process() allow to have a more consistent
system but is not sufficient. In fact, transaction should be limited to
one HTTP round trip to prestashop in an ideal world.