expense: fix translation handling
All translated fields where checking row['message_key'] != "" whereas
its value is null when no translation is present. This was causing
missing elements in the report, especially the "kilometric allow param",
when in foreign language (eg. french).
Moreover fix a typo in Kilometric Allowance preventing the title to be
properly printed on report.
The kilometric_allow_param can be translated in the report which is
strange since these are user defined strings, but leave it this way as
changing it might trigger https://xkcd.com/1172/