account: add an option to allow editing of validated move
Add a new option to allow editing validated account moves as long as
they haven't been exported. This allows to be more flexible when an
external accouting software is used.
This works around the fact that the accountingDayBook option is almost
never used (and the code flow in the invoicing system is such a maze
that this can not be fixed on a short term).