reports: add some utility classes for JasperReports
Add a custom processor for HTML markup to get closer from the BIRT
rendering wrt line feeds & lists. This has to be enhanced as more and
more reports are converted.
Add a static method to run a query against a connection in order to
avoid using lists. This is not perfect as we lose type information but
can help to convert quicker from BIRT.