fixedAssets: create from invoice when category is filled
Do not check whether config flag fixedAssetCatReqOnInvoice is set to
create the fixed asset. If user filled the category it is likely that he
wants to create the asset. Moreover the label of the option does not
indicate in any way that setting it to false will prevent any automatic
fixed asset creation.