taxEquiv: fix reverse charge tax computation
Reverse charge computation was broken when more than one product was on
a purchase order or on an invoice.
Reverse tax charge was initialized with first line ex tax amount and
subsequent lines' ex tax amount were subtracted instead of added. This
is clearly an inconsistency since the final computation is done by
negating the total ex tax.