Compute future qty in location line from a query
Future qty in location lines was a counter, so if a bug happened the
value stored in location lines was not coherent with values in stock
move lines.
This commit allows to compute the future qty from planned stock moves,
so if an error happens it will be fixed on the next stock move plan.