RM18571: new boolean qtyRequested in so line
If qtyRequested is true, then we request reservation for the line
(meaning that qtyRequested will be equal to the quantity of the line).

This fix issues with the requested reserved qty in sale order line,
which has to take into account delivered quantity, that can be confusing
for the user. With this commit we simply show the user whether the line
is requesting a reservation or not.