Signatures in user & template
User's new field: emailSignature.
On templates (from message) add a new formula field: the email signature.
This formula field may point on a user's emailSignature directly or from an emailAccount's signature.

When an email is created from template, the signature is automaticly add at the end of the email's content.
If the formula field was empty or incorrectly filled, the default emailAccount's signature is automaticly add at the end of the mail instead.

RM#26313 (main features commit, missing the autofill of the formula field)