CostSheet: improve price computation from WAP
When we use WAP to get the cost of a component, sometimes we cannot have
its correct WAP per company because there are no quantities in stock.
Now when this happens, we get the average price between existing wap
price in stock location linked to the correct company.
Fixes RM27102