Configurator: add free field in indicator list
Refactor the configurator creator model so it has only one one-to-many,
and remove different classes of configurator formula lines. We now check
if the line is for a product or for a sale order line using a new field,
typeSelect. A third value was added for typeSelect to allow a new type
of line: a free field which does not need to be an existing meta field
in sale order line or product.

RM36346